refactor: technical audit — security, dedup, CI/CD
- Dockerfiles: use built-in non-root app user (.NET 10 images)
- Fix swallowed exceptions in UserBookService, PdfTextExtractor
- Search query max length 200 chars (ReDoS prevention)
- Pin LibreTranslate v1.6.4, remove default DB password fallback
- Startup config validation (JWT_SECRET, GOOGLE_CLIENT_ID)
- CORS origins from appsettings, health check w/ DB query
- Extract ImageProcessingHelper (dedup IngestionService×2)
- Extract GetUserId extension (dedup 4 endpoint files)
- Consolidate authFetch in frontend client.ts (dedup 4 modules)
- Add Result<T>, JsonDefaults shared options
- CI: add dotnet format lint step, SSG page count validation
- Makefile: add test/lint targets
- FileUpload size limits in appsettings.json
