+++ Makefile
@@ -1,4 +1,4 @@
-.PHONY: up down restart logs status backup restore rebuild-ssg clean-ssg deploy nginx-setup build rebuild fix-permissions
+.PHONY: up down restart logs status backup restore rebuild-ssg clean-ssg deploy nginx-setup build rebuild fix-permissions test lint
 
 # ============================================================
 # Docker Services
@@ -75,6 +75,17 @@ rebuild-ssg:
 clean-ssg:
 	rm -rf apps/web/dist/ssg apps/web/dist/ssg-new apps/web/dist/ssg-old
 	@echo "SSG cleaned"
+
+# ============================================================
+# Testing & Linting
+# ============================================================
+
+test:
+	dotnet test
+	pnpm -C apps/web test
+
+lint:
+	dotnet format --verify-no-changes
 
 # ============================================================
 # Nginx Setup (one-time)
