+++ apps/web/src/api/userBooks.ts
@@ -1,6 +1,4 @@
-import { refreshToken } from './auth'
-
-const API_BASE = import.meta.env.VITE_API_URL ?? ''
+import { authFetch, API_BASE } from './client'
 
 export interface UserBook {
   id: string
@@ -65,37 +63,6 @@ export interface StorageQuota {
   usedBytes: number
   limitBytes: number
   usedPercent: number
-}
-
-async function authFetch<T>(path: string, options?: RequestInit, retry = true): Promise<T> {
-  const res = await fetch(`${API_BASE}${path}`, {
-    ...options,
-    credentials: 'include',
-  })
-
-  if (!res.ok) {
-    if (res.status === 401 && retry) {
-      // Try refresh token and retry once
-      try {
-        await refreshToken()
-        return authFetch<T>(path, options, false)
-      } catch {
-        throw new Error('Unauthorized')
-      }
-    }
-    if (res.status === 401) throw new Error('Unauthorized')
-    const text = await res.text()
-    let error = `API error: ${res.status}`
-    try {
-      const json = JSON.parse(text)
-      if (json.error) error = json.error
-    } catch {}
-    throw new Error(error)
-  }
-
-  const text = await res.text()
-  if (!text) return {} as T
-  return JSON.parse(text)
 }
 
 export async function uploadUserBook(
