+++ backend/Docker/Api.Dockerfile
@@ -19,6 +19,7 @@ FROM mcr.microsoft.com/dotnet/aspnet:10.0-alpine AS runtime
 RUN apk add --no-cache git
 WORKDIR /app
 COPY --from=build /app/publish .
+USER app
 ENV ASPNETCORE_URLS=http://+:8080
 EXPOSE 8080
 ENTRYPOINT ["dotnet", "Api.dll"]
