+++ backend/src/Api/appsettings.json
@@ -23,5 +23,22 @@
     "BaseUrl": "http://localhost:5000",
     "MaxTextLength": 500,
     "TimeoutSeconds": 30
+  },
+  "FileUpload": {
+    "MaxBookSizeMb": 100,
+    "MaxCoverSizeMb": 5,
+    "MaxPhotoSizeMb": 2
+  },
+  "Cors": {
+    "AllowedOrigins": [
+      "http://localhost:5173",
+      "http://general.localhost",
+      "http://general.localhost:5173",
+      "http://localhost:81",
+      "http://admin.localhost",
+      "http://admin.localhost:81",
+      "https://textstack.app",
+      "https://textstack.dev"
+    ]
   }
 }
