+++ backend/src/Application/Common/JsonDefaults.cs
@@ -1,0 +1,11 @@
+using System.Text.Json;
+
+namespace Application.Common;
+
+public static class JsonDefaults
+{
+    public static readonly JsonSerializerOptions Options = new()
+    {
+        PropertyNameCaseInsensitive = true
+    };
+}
