+++ backend/src/Worker/Services/UserIngestionService.cs
@@ -3,7 +3,6 @@ using Application.Common.Interfaces;
 using Domain.Entities;
 using Domain.Enums;
 using Domain.Utilities;
-using HtmlAgilityPack;
 using Infrastructure.Persistence;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
@@ -183,7 +182,7 @@ public class UserIngestionService
             // Create chapters
             foreach (var unit in result.Units)
             {
-                var html = RewriteImageSrcs(unit.Html ?? string.Empty, imageMap);
+                var html = Application.Common.ImageProcessingHelper.RewriteImageSrcs(unit.Html ?? string.Empty, imageMap);
                 var chapterTitle = SanitizeText(unit.Title ?? $"Chapter {unit.OrderIndex + 1}");
                 var chapter = new UserChapter
                 {
@@ -251,64 +250,6 @@ public class UserIngestionService
         }
     }
 
-    private static string RewriteImageSrcs(string html, Dictionary<string, string> imageMap)
-    {
-        if (string.IsNullOrEmpty(html) || imageMap.Count == 0)
-            return html;
-
-        var doc = new HtmlDocument();
-        doc.LoadHtml(html);
-
-        var imgNodes = doc.DocumentNode.SelectNodes("//img[@src]");
-        if (imgNodes == null)
-            return html;
-
-        foreach (var img in imgNodes)
-        {
-            var src = img.GetAttributeValue("src", "");
-            if (string.IsNullOrEmpty(src))
-                continue;
-
-            var normalizedSrc = NormalizeImagePath(src);
-
-            foreach (var (originalPath, newUrl) in imageMap)
-            {
-                var normalizedOriginal = NormalizeImagePath(originalPath);
-                if (normalizedSrc.Equals(normalizedOriginal, StringComparison.OrdinalIgnoreCase) ||
-                    normalizedSrc.EndsWith(normalizedOriginal, StringComparison.OrdinalIgnoreCase) ||
-                    normalizedOriginal.EndsWith(normalizedSrc, StringComparison.OrdinalIgnoreCase))
-                {
-                    img.SetAttributeValue("src", newUrl);
-                    break;
-                }
-            }
-        }
-
-        return doc.DocumentNode.InnerHtml;
-    }
-
-    private static string NormalizeImagePath(string path)
-    {
-        var result = path;
-        while (result.StartsWith("../"))
-            result = result[3..];
-        while (result.StartsWith("./"))
-            result = result[2..];
-        result = result.TrimStart('/');
-        return result;
-    }
-
-    private static string GetExtensionFromMimeType(string mimeType)
-    {
-        return mimeType switch
-        {
-            "image/png" => ".png",
-            "image/gif" => ".gif",
-            "image/webp" => ".webp",
-            "image/svg+xml" => ".svg",
-            _ => ".jpg"
-        };
-    }
 
     private static string MapToFriendlyError(ExtractionWarningCode? code) => code switch
     {
